% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{animator}
\alias{animator}
\alias{as.animator}
\alias{is.animator}
\title{Plot Block of Animated Commands.}
\usage{
animator(x, life = 1, fps = 25, pause = 0.5, verbose = FALSE)

as.animator(x, life = 1)

is.animator(x)
}
\arguments{
\item{x}{character, expression (block) or
object of class \code{animator} containing  graphical timed commands.}

\item{life}{numerical, duration of animation.}

\item{fps}{numerical, frames per second.}

\item{pause}{numerical, length of the pause between plotted frames.}

\item{verbose}{logical, if TRUE print animation characteristics.}
}
\value{
Function \code{animator} plots the animation and
  returns an object of class \code{animator}.

 \code{as.animator} attempts to add the class
 \code{animator} to the argument \code{x}.

 \code{is.animator} returns \code{TRUE} if \code{x}
 is an \code{animator} object and \code{FALSE} otherwise.
}
\description{
Make and plot animated sequence of figures.
}
\examples{
if(interactive())
animator("newplot();tpoints(2,2,5,8,cex=2,pch=16)",life=2,verbose=TRUE)
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\seealso{
\code{\link{as.animator}}, \code{\link{is.animator}}.
}

