% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{as.animator}
\alias{as.animator}
\title{Objects of class \code{animator}.}
\usage{
as.animator(x, life = 1)
}
\description{
Creates or tests for objects of  class \code{animator}.
}
\examples{
x <- as.animator(
"newplot();tpoints(2,2,5,8,cex=2,pch=16)", life=2)
print(x)
if(interactive()) plot(x)
## Equivalent
x <- as.animator(expression({
newplot()
tpoints(2,2,5,8,cex=2,pch=16)
}),
life=2)
print(x)
if(interactive()) plot(x)
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\seealso{
\code{\link{animator}}.
}

