% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Combine multiple dataframes with different columns together}
\usage{
combine(...)
}
\arguments{
\item{...}{One or more dataframes}
}
\value{
A dataframe with all of the rows from the original dataframes
appended onto each other and missing column values filled in with NAs.
}
\description{
Combine multiple dataframes with different columns together
}
\details{
This function is similar to \code{rbind()} in that multiple dataframes
can be appended on top of each other. Unlike \code{rbind()} though, the dataframes
do not have to have the same columns. A dataframe that is missing a column
that another dataframe has will have its portion of the final output filled
in with NAs.

This function can be used when series of actions were created in multiple
dataframes and some are relying on default values and don't contain the
same columns as others.
}
