% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Convert dataframes of actions into an web page with the animation}
\usage{
compile(
  path = ".",
  bg,
  warn = TRUE,
  images_df = NULL,
  text_df = NULL,
  size = NULL
)
}
\arguments{
\item{path}{The relative or absolute path to the location where a new folder
will be created called animation_output containing the files necessary
to run the animation.}

\item{bg}{The relative or absolute file path to the image that will be used
as the background.}

\item{warn}{If set to TRUE and there is already a folder in the given path
called "animation_output", this will prompt you to confirm whether it should
be overwritten. Set to FALSE if you always want it to automatically overwrite
any old animations.}

\item{images_df}{A dataframe in the structure provided by images_template()}

\item{text_df}{A dataframe in the structure provided by text_template()}

\item{size}{The dimensions of the output. If left as NULL, the size of the
background image will be used.}
}
\description{
Convert dataframes of actions into an web page with the animation
}
