% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{coordParams}
\alias{coordParams}
\title{Coordinate Arguments}
\arguments{
\item{x0}{numeric vector, start x coordinates.}

\item{y0}{numeric vector, start y coordinates.}

\item{x1}{numeric vector, end x coordinates.}

\item{y1}{numeric vector, end y coordinates.}
}
\description{
The core coordinate arguments for animatoR plotting functions
}

