% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{getChunkopts}
\alias{getChunkopts}
\title{Get Chunk Options.}
\usage{
getChunkopts(what)
}
\arguments{
\item{what}{character, name of the options to extract.
If missing, all options will be returned.}
}
\value{
\code{Sweave} or \code{knitr} chunk option with name
  as declared by argument \code{what}.
  If \code{what} is missing, a list with all chunk options.
}
\description{
Get \code{Sweave} or \code{knitr} programme chunk options.
}
\examples{
\dontrun{
getChunkopts()
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}

