% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{images_template}
\alias{images_template}
\title{Create an empty dataframe that contains the columns needed for the \code{images_df}
argument in the \code{compile()} function.}
\usage{
images_template()
}
\description{
Create an empty dataframe that contains the columns needed for the \code{images_df}
argument in the \code{compile()} function.
}
\details{
The columns that are compatible in images dataframes are listed below.
Any that are marked as optional can either contain NAs to use their default
value, or the whole column can be omitted if the default behavior is desired
for every action.
\itemize{
\item \strong{startTime} (numeric) - How many seconds into the animation that the image should
start to appear.
\item \strong{endTime} (numeric) - How many seconds into the animation that the image should
disappear.
\item \strong{startX} (numeric) - The X coordinate where the center of the image will
begin.
\item \strong{startY} (numeric) - The Y coordinate where the center of the image will
begin.
\item \strong{endX} (numeric) - The X coordinate where the center of the image will
end.
\item \strong{endY} (numeric) - The Y coordinate where the center of the image will
end.
\item \strong{startWidth} (numeric) - How many pixels wide the image should initially be.
\item \strong{startHeight} (numeric) - How many pixels tall the image should initially be.
\item \strong{endWidth} (numeric, optional) - How many pixels wide the image should be
by the endTime. If set as NA or if the column is not provided, the image's
width will not change.
\item \strong{endHeight} (numeric, optional) - How many pixels tall the image should be
by the endTime. If set as NA or if the column is not provided, the image's width
will not change.
\item \strong{startOpacity} (numeric, optional) - How opaque the image should be when it
first appears, between 0 and 1. 0 is completely transparent and 1 is completely
opaque. If set as NA or if the column is not provided, the image will be completely
opaque.
\item \strong{endOpacity} (numeric, optional) - How opaque the image should be by its
endTime, between 0 and 1. 0 is completely transparent and 1 is completely
opaque. If set as NA or if the column is not provided, the opacity of the image
will not change.
\item \strong{image} (character) - An absolute or relative path to an image file (png,
jpg, gif).
}
}
