% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{includeLatex}
\alias{includeLatex}
\title{Include Animated Graphics.}
\usage{
includeLatex(title = "", file = NA, scale = 0.5, poster = "first",
  every = 1, fps = 25, first = "", last = "", vspace = "0pt",
  other = "controls")
}
\arguments{
\item{title}{character, obsolete.}

\item{file}{character,  the leftmost part of the file name that
is common to all members of the sequence. If \code{file} is
equal to \code{NA}, \code{file} will be constructed from the
code graphics prefix and chunk label. See Note.}

\item{scale}{numeric, scaling factor, See Note.}

\item{poster}{character or numeric,
['first' | <num> | 'last' | 'none']
Specifies which frame to display and print if
the animation is not activated. The first frame is shown by default.
Thus ‘poster’ or ‘poster=first’ need not be explicitly set.
A frame number <num> may as well be given; <num> is zero-based,
that is, the first frame has number ‘0’. See Note.}

\item{every}{numeric. Build animation from every 'every'th frame only.
Skipped frames are discarded and not embedded into the document.
See Note.}

\item{fps}{numeric,  the animation frame rate (frames per second).
Assigning values less than zero results in an error.
Default is 25. See Note.}

\item{first}{numeric, first frame to use in animation.
Special case is default character value "",
equivalent to 1. See Note.}

\item{last}{numeric, last frame to use in animation.
Special case is default character value "",
equivalent to last, unknown frame number. See Note.}

\item{vspace}{character, vertical space (LaTeX style)
for graph positioning. Default is "0pt", no additional space.}

\item{other}{character, other parameters separated by coma and
passed to LaTeX package \code{animate} See Note.}
}
\value{
prints and invisibly returns the
  LaTeX \code{\\animategraphics} command.
}
\description{
Provides a line to embed a stack of graphic frames
into PDF file
as animated graphics. It is an interface for LaTeX command
\code{\\animategraphics} from LaTeX package \code{animate}.
}
\note{
Arguments (except \code{title} and
  \code{vspace}) are used to communicate
  with LaTeX package \emph{animate}.
  For details see documentation for the LaTeX package
  \href{https://www.ctan.org/pkg/animate}{animate}.
  Be aware that this is not the same as R package
  \code{\link[animation]{animation}}.

  When used with Sweave or knitr, the argument \code{file}
  need not to be set and will be conveniently constructed from the
  graphical prefix and chunk label.

  Argument \code{vspace} can be useful for raising the
  animation image on slides (e.g. \code{beamer}).
}
\examples{
includeLatex("Test animation")
includeLatex("Test animation",scale=0.5,poster="last",other="loop")
# To include from specific file with stacked frames ( e. g. 0 .. 99 )
# Useful to include animations that were prepared before and can
# be reused in another file
includeLatex(file="./figs/PreparedBefore",first=10,last=50)
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\seealso{
\code{\link{animator}},
  LaTeX package \code{animate} (\url{https://www.ctan.org/pkg/animate}),
  R package \code{animation}
  (\url{https://cran.r-project.org/web/packages/animation)}.
}

