% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{is_knitr}
\alias{is_knitr}
\title{Test if \code{knitr} is Active}
\usage{
is_knitr()
}
\value{
value \code{TRUE} if \code{knitr} engine is active,
  \code{FALSE} otherwise
}
\description{
Are you \code{knitr}?
}
\examples{
\dontrun{
is_knitr()
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}

