% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{makeTransparent}
\alias{makeTransparent}
\title{Make Transparent Colors}
\usage{
makeTransparent(x, alpha = 0.5)
}
\arguments{
\item{x}{vector specifying colors or factor.
Colors can be specified as numbers or
character strings, see \code{\link{colors}}.
Factors are also acceptable
in which case the input will be transformed into level numbers.}

\item{alpha}{numeric, transparency value \code{\link{rgb}}.}
}
\value{
A character vector with elements of 7 or 9 characters, "#"
  followed by the red, blue, green and optionally alpha
  values in hexadecimal (after rescaling to 0 ... 255).
  The optional alpha values range from 0
  (fully transparent) to 255 (opaque).
}
\description{
Add transparency parameter (alpha) to colors
}
\examples{
# Transparent red
makeTransparent("red",0.5)
# Same thing
makeTransparent(2,0.5)
# Vector of names
makeTransparent(c("red","orange"),0.75)
# Vector of color numbers
makeTransparent(1:5)
# Factor levels can be used as color indicator
makeTransparent(factor(c("S","M","H","S")))
# Different transparency
makeTransparent(1:5,alpha=1:5)
#
# Define transparent version of base colors
#
trcols<-makeTransparent(c(1:8,0.1),alpha=0.75)
trcols <- data.frame(t(trcols),stringsAsFactors=FALSE)
names(trcols) <-
paste("tr",c("black","red","green","blue",
"cyan","magenta","yellow","grey","white"),sep="")
str(trcols)
attach(trcols)
trred
trgreen
detach()
# Plot colors
if(interactive()){
par(xpd=TRUE)
newplot(stamp=FALSE)
points(1:9,1:9,bg=unlist(trcols),col="grey",pch=21,cex=15)
text((1:9)+1.1,(1:9)-0.7,names(trcols),adj=0)
}

}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\seealso{
\code{\link[grDevices]{rgb}} for setting colors and
  \code{\link[grDevices]{colors}} for color names.
}
\keyword{aplot}
\keyword{dynamic}
\keyword{graphics}

