% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_defaults.R
\name{replace_nas_with_defaults}
\alias{replace_nas_with_defaults}
\title{Replace NA values in optional columns with default values}
\usage{
replace_nas_with_defaults(df)
}
\arguments{
\item{df}{Either an images or text dataframe}
}
\value{
A modified version of the provided df dataframe whose NA values in
the optional columns have been replaced.
}
\description{
Replace NA values in optional columns with default values
}
