% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{tBoxCox}
\alias{tBoxCox}
\title{Box-Cox Transformation with Interpolated Parameter \eqn{lambda}}
\usage{
tBoxCox(x, lambda0, lambda1 = lambda0, t, when, p)
}
\arguments{
\item{x}{numerical vector.}

\item{lambda0}{numeric, starting value of parameter lambda.}

\item{lambda1}{numeric, end value of parameter lambda.}

\item{t}{numeric, homotopy parameter, limited between 0 and 1.
This parameter can be considered as fraction of animation duration time.}

\item{when}{numeric vector. This parameter controls
the times of: entrance, exit, start of movement and, end of movement.}

\item{p}{numeric, homotopy power parameter. Defaults to 1.}
}
\value{
numerical vector, Box-Cox transformed
  values for current interpolated value of lambda.
}
\description{
Transform data using the Box-Cox transformation with
parameter \eqn{lambda} interpolated between start and end value.
}
\examples{
if(interactive())
x <- runif(250,1,3)
animator("newplot(xlim=c(0,max(x)^2),ylim=c(0,1),axes=TRUE,asp=NA);
xt <- tBoxCox(x,-1,3,t); rug(xt);
lines(density(xt,add=TRUE))",life=5,verbose=TRUE)
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}

