% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{tcex}
\alias{tcex}
\title{Interpolate Symbol and Text Size}
\usage{
tcex(cex0 = 1, cex1 = 1, t, when, p, area = TRUE)
}
\arguments{
\item{t}{numeric, homotopy parameter, limited between 0 and 1.
This parameter can be considered as fraction of time
duration of the animation.}

\item{when}{numeric vector. This parameter controls
the times of: entrance, exit, start of movement and, end of movement.}

\item{p}{numeric, homotopy power parameter. Defaults to 1.}

\item{x0}{numeric vector specifying start symbol size,
see \code{\link[graphics]{par} cex}.}

\item{x1}{numeric vector specifying end symbol size,
see \code{\link[graphics]{par} cex}.}
}
\value{
A numeric vector giving the amount by which plotting text
and symbols should be magnified relative to the default \code{cex} value.
}
\description{
This function interpolates the symbol size
between the start and end size.
}
\examples{
if(interactive()){
animator('
newplot()
points(3:7,3:7,col=1:5,
pch=16,cex=tcex(0,20,t),life=0.5)
')
}
if(interactive()){
# Doubling the symbol sizes
newplot()
text(2,8,"By edge:\\n cex = c(8, 16)")
points(2,5,cex=16,pch=0)
points(2-0.75*par("cin")[2]*4,
5-0.75*par("cin")[2]*4,cex=8,pch=15, col="grey")
text(8,8,"By size:\\n cex = sqrt(c(4, 16, 64))")
points(8,5,cex=sqrt(64*4),pch=0)
points(8,5,cex=sqrt(16*4),pch=15, col="grey")
points(8,5,cex=sqrt(4*4),pch=0)
abline(v=c(2,8),h=5,col="red")
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\seealso{
\code{\link[graphics]{par}} for setting symbol sizes.
}
\keyword{aplot}
\keyword{dynamic,}

