% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{text_template}
\alias{text_template}
\title{Create an empty dataframe that contains the columns needed for the \code{text_df}
argument in the \code{compile()} function.}
\usage{
text_template()
}
\description{
Create an empty dataframe that contains the columns needed for the \code{text_df}
argument in the \code{compile()} function.
}
\details{
The columns that are compatible in text dataframes are listed below.
Any that are marked as optional can either contain NAs to use their default
value, or the whole column can be omitted if the default behavior is desired
for every action.
\itemize{
\item \strong{startTime} (numeric) - How many seconds into the animation that the text should
start to appear.
\item \strong{endTime} (numeric) - How many seconds into the animation that the text should
disappear.
\item \strong{startX} (numeric) - The X coordinate where the left side of the text will
begin.
\item \strong{startY} (numeric) - The Y coordinate where the center of the text will
begin.
\item \strong{endX} (numeric) - The X coordinate where the left of the text will
end.
\item \strong{endY} (numeric) - The Y coordinate where the center of the text will
end.
\item \strong{font} (character) - A string in the form of an \href{https://www.w3schools.com/tags/canvas_font.asp}{HTML canvas's font property}.
For example, "bold 12px arial".
\item \strong{fillStyle} (character, optional) - The color of the text. Can either be a \href{https://www.w3schools.com/tags/ref_colornames.asp}{color name}
supported by your internet browser such as "green" or a hexcode such as "#FFD700".
If set to NA or if the column is not provided, the text will be black.
\item \strong{text} (character) - The text to display.
}
}
