% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{tlines}
\alias{tlines}
\title{Move Lines}
\usage{
tlines(x0, y0, x1, y1, t, p = 1, when, ...)
}
\arguments{
\item{x0}{numeric vector, start x coordinates.}

\item{y0}{numeric vector, start y coordinates.}

\item{x1}{numeric vector, end x coordinates.}

\item{y1}{numeric vector, end y coordinates.}

\item{t}{numeric, homotopy parameter, limited between 0 and 1.
This parameter can be considered as fraction of animation duration time.}

\item{p}{numeric, homotopy power parameter. Defaults to 1.}

\item{when}{numeric vector. This parameter controls
the times of: entrance, exit, start of movement and, end of movement.}

\item{...}{additional arguments passed to \code{\link{lines}}.}
}
\value{
List with numerical components \code{x} and \code{y} with
  current position
}
\description{
Move lines from start to end location
}
\examples{
x0 <- c(0,5,10)
y0 <- c(0,10,0)
x1 <- c(10,5,0)
y1 <- c(10,0,10)
newplot()
pos <- tlines(x0,y0,x1,y1,0.25)
points(pos)
#############
par(mfrow=c(2,2))
for( t in seq(0,1,1/3)) {
newplot()
tlines(x0,y0,x1,y1,t)
points(x0,y0,pch="0")
points(x1,y1,pch="1")
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\seealso{
\code{\link[graphics]{lines}}
}
\keyword{aplot}
\keyword{dynamic}
\keyword{graphics}

