% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{tmatrix}
\alias{tmatrix}
\title{Interpolate Matrix}
\usage{
tmatrix(X0, X1 = diag(nrow(X0)), t, when, p)
}
\arguments{
\item{t}{numeric, homotopy parameter, limited between 0 and 1.
This parameter can be considered as fraction of time
duration of the animation.}

\item{when}{numeric vector. This parameter controls
the times of: entrance, exit, start of movement and, end of movement.}

\item{p}{numeric, homotopy power parameter. Defaults to 1.}
}
\value{
Intermediate matrix (numerical).
}
\description{
Interpolates matrix elements between start and end value.
}
\examples{
X0 <- matrix(c(1,2,3,4),2,2)
X1 <- matrix(c(11,12,13,14),2,2)                     
print(X0)                                            
print(X1)                                            
print(h(X0,X1,t=0.3))                    
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}

