% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{tParams}
\alias{tParams}
\title{Argument t}
\arguments{
\item{t}{numeric, homotopy parameter, limited between 0 and 1.
This parameter can be considered as fraction of time
duration of the animation.}
}
\description{
Homotopy argument
}

