% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{tpoints}
\alias{tpoints}
\title{Move points}
\usage{
tpoints(x0, y0, x1 = x0, y1 = y0, t, when = c(0, 1), trace = FALSE,
  trace.col = "grey", ...)
}
\arguments{
\item{x0}{numeric vector, start x coordinates.}

\item{y0}{numeric vector, start y coordinates.}

\item{x1}{numeric vector, end x coordinates.}

\item{y1}{numeric vector, end y coordinates.}

\item{t}{numeric, homotopy parameter, limited between 0 and 1.
This parameter can be considered as fraction of time
duration of the animation.}

\item{when}{numeric vector. This parameter controls
the times of: entrance, exit, start of movement and, end of movement.}

\item{trace}{logical. Should movement leave a trace?}

\item{trace.col}{color of the trace.}

\item{...}{other parameters passed to \code{\link{points}}.}
}
\value{
List with numerical components \code{x} and \code{y} with
current position
}
\description{
Move points from start to end location
}
\examples{
x0 <- c(0,5)
y0 <- c(0,5)
x1 <- c(1,10)
y1 <- c(1,10)
print(tpoints(x0,y0,x1,y1,0.5))
#############
par(mfrow=c(2,2))
for( t in seq(0,1,1/3)) {
newplot()
tpoints(x0,y0,x1,y1,t,trace=TRUE)
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\seealso{
\code{\link[graphics]{points}}
}
\keyword{aplot}
\keyword{dynamic,}

