% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{tsegments}
\alias{tsegments}
\title{Move Segments}
\usage{
tsegments(x0, y0, x1, y1, x2 = x0, y2 = y0, x3 = x1, y3 = y1, t, when,
  p = 1, fixed = 0, ...)
}
\arguments{
\item{x0, x1}{numeric vectors, start segments defining x coordinates.}

\item{y0, y1}{numeric vector, start segments defining y coordinates.}

\item{x2, x3}{numeric vector, end segments defining x coordinates.}

\item{y2, y3}{numeric vector, end segments defining y coordinates.}

\item{t}{numeric, homotopy parameter, limited between 0 and 1.
This parameter can be considered as fraction of animation duration time.}

\item{when}{numeric vector. This parameter controls
the times of: entrance, exit, start of movement and, end of movement.}

\item{p}{numeric, homotopy power parameter. Defaults to 1.}

\item{fixed}{numeric, which location is fixed:
start (0, default) - draw from \code{(x0,y0)} to \code{(x1,y1)},
or end (1) - reverse.}

\item{...}{additional arguments passed to \code{\link{segments}}.}
}
\value{
Numerical matrix with columns defining the current segments.
}
\description{
Change segment defining positions and plot subsequent segments.
Lines are "floating" to final positions.
}
\examples{
x0 <- c(0,4,7)
y0 <- c(0,7,0)
x1 <- c(4,0,7)
y1 <- c(2,7,5)
x2 <- x0+1
y2 <- y0+2
x3 <- x1+1
y3 <- y1+2
newplot()
arrows(x0,y0,x2,y2,lty=2)
arrows(x1,y1,x3,y3,lty=2)
segments(x0,y0,x1,y1,col=2)
pos <- tsegments(x0,y0,x1,y1,x2,y2,x3,y3,0.75,lwd=3)
segments(x2,y2,x3,y3,col=4)
pos
points(pos$start)
points(pos$end)
d <- 0.2
points(x0-d,y0+d,pch="0")
points(x1+d,y1+d,pch="1")
points(x2-d,y2+d,pch="2")
points(x3+d,y3+d,pch="3")
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\seealso{
\code{\link[graphics]{segments}}, \code{\link{dsegments}}
  for drawing segments and \code{\link{tarrows}} for moving arrows.
}
\keyword{aplot}
\keyword{dynamic}
\keyword{graphics}

