% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{ttext}
\alias{ttext}
\title{Move Text}
\usage{
ttext(x0, y0, x1 = x0, y1 = y0, t, when, text = "", ...)
}
\arguments{
\item{x0}{numeric vector, start x coordinates.}

\item{y0}{numeric vector, start y coordinates.}

\item{x1}{numeric vector, end x coordinates.}

\item{y1}{numeric vector, end y coordinates.}

\item{t}{numeric, homotopy parameter, limited between 0 and 1.
This parameter can be considered as fraction of animation duration time.}

\item{when}{numeric vector. This parameter controls
the times of: entrance, exit, start of movement and, end of movement.}

\item{text}{character vector, tect to display.}

\item{...}{additional arguments passed to \code{\link{text}}.}
}
\value{
List with numerical components \code{x} and \code{y} with
  current position.
}
\description{
Move text from start to end location
}
\examples{
x0 <- c(0,5)
y0 <- c(0,5)
x1 <- c(1,10)
y1 <- c(1,10)
print(tpoints(x0,y0,x1,y1,0.5))
#############
par(mfrow=c(2,2))
for( t in seq(0,1,1/3)) {
newplot()
ttext(x0,y0,x1,y1,t,text=paste("Test text",t))
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\seealso{
\code{\link[graphics]{text}}
}
\keyword{aplot}
\keyword{dynamic}
\keyword{graphics}

