% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_compile_input.R
\name{validate_compile_input}
\alias{validate_compile_input}
\title{Basic checks to see if there are any simple problems that would cause the compile() function to crash}
\usage{
validate_compile_input(path, bg, warn, images_df, text_df, size)
}
\description{
Basic checks to see if there are any simple problems that would cause the compile() function to crash
}
\details{
All arguments passed to this function should match the arguments passed to compile()
}
