% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{whenParams}
\alias{whenParams}
\title{Argument when}
\arguments{
\item{when}{numeric vector. This parameter controls
the times of: entrance, exit, start of movement and, end of movement.}
}
\description{
Description of argument when
}

