% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{animator}
\alias{animator}
\title{Plot Animated Block of Commands.}
\usage{
animator(block, life = 1, fps = 25, pause = 0.5, verbose = FALSE)
}
\value{
object of class \code{animator}
}
\description{
Main function that plots animated sequence of figures.
}
\examples{
if(interactive())
animator("newplot();tpoints(2,2,5,8,cex=2,pch=16)",life=2,verbose=TRUE)
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}

