% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\docType{package}
\name{animatoR}
\alias{animatoR}
\alias{animatoR-package}
\title{animatoR: Animated Graphics in R (Base Graphics)}
\description{
The package provides several functions for support
of animated graphics in base R graphics.
}
\section{Usage}{


Animated graphics can be useful for demonstration of
results that change in a linear succession
(e.g. time or direction).
It can be used for making of statistical 'cartoons'
for demonstartion of statistical concepts in teaching.
}

