% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{dsegments}
\alias{dsegments}
\title{Draw Segments}
\usage{
dsegments(x0, y0, x1, y1, t, when, p = 1, ...)
}
\arguments{
\item{x0}{numeric vector, start x coordinates.}

\item{y0}{numeric vector, start y coordinates.}

\item{x1}{numeric vector, end x coordinates.}

\item{y1}{numeric vector, end y coordinates.}

\item{t}{numeric, homotopy parameter, limited between 0 and 1.
This parameter can be considered as fraction of time
duration of the animation.}

\item{when}{numeric vector. This parameter controls
the times of: entrance, exit, start of movement and, end of movement.}

\item{p}{numeric, homotopy power parameter. Defaults to 1.}

\item{...}{other parameters passed to \code{\link{segments}}.}
}
\value{
List with numerical components \code{x} and \code{y} with
current position
}
\description{
Draw segments from (x0,y0) to (x1,y1). The effect is
like starting from a point and draw a line.
}
\examples{
x0 <- c(0,5,10)
y0 <- c(0,10,0)
x1 <- c(10,5,0)
y1 <- c(10,0,10)
newplot()
pos <- dsegments(x0,y0,x1,y1,0.25)
points(pos)
## Lines from 0 to 1
par(mfrow=c(2,2))
for( t in seq(0,1,1/3)) {
newplot()
dsegments(x0,y0,x1,y1,t)
points(x0,y0,pch="0")
points(x1,y1,pch="1")
}
## Reverse: from 1 to 0
par(mfrow=c(2,2))
for( t in seq(0,1,1/3)) {
newplot()
dsegments(x1,y1,x0,y0,t)
points(x0,y0,pch="0")
points(x1,y1,pch="1")
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\seealso{
\code{\link[graphics]{segments}}
}
\keyword{graphics}

