% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{h}
\alias{h}
\title{Homotopy Function}
\usage{
h(x0, x1, t, when, p = 1)
}
\arguments{
\item{x0}{numeric vector of start values.}

\item{x1}{numeric vector of end values.}

\item{t}{numeric, homotopy parameter, limited between 0 and 1.
This parameter can be considered as fraction of animation duration time.}

\item{when}{numeric vector. This parameter controls
the times of: entrance, exit, start of movement and, end of movement.}

\item{p}{numeric, homotopy power parameter. Defaults to 1.}
}
\value{
interpolated value (see Note)
}
\description{
Interpolates a position between start and end value(s).
Homotopy is controled by a homotopy parameter \code{t} and power parameter \code{p} is used for interpolation.
}
\note{
Returned coordinates are determined using the homotopy function
  \deqn{x_t=x_0  (1-t^p)+x_1 t^p,  t\in[0,1]}
}
\examples{
h(1, 2, 0)
h(1, 2, 1)
h(1, 2, .5)
h(1,c(2,3),.5)

}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{dynamic}

