% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{newplot}
\alias{newplot}
\title{Open New Empty Plotting Window}
\usage{
newplot(xlim = c(0, 10), ylim = c(0, 10), ann = FALSE, axes = FALSE,
  type = "n", asp = 1, stamp = FALSE, ...)
}
\arguments{
\item{xlim}{see \code{\link{plot}}}

\item{ylim}{see \code{\link{plot}}}

\item{ann}{see \code{\link{plot}}}

\item{axes}{see \code{\link{plot}}}

\item{type}{see \code{\link{plot}}}

\item{asp}{aspect ratio, default y/x = 1 see \code{\link{plot.new}}}

\item{stamp}{logical, should time be visible on plots.}

\item{...}{see \code{\link{plot}}}
}
\description{
Opens a new empty plotting window with default settings
 (0, 10) x (0, 10), no axes nor labels.
}
\examples{
newplot()
points(0.5,0.5)

}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\seealso{
\code{\link{plot}}
}
\keyword{dynamic,}
\keyword{hplot}

