% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{tarrows}
\alias{tarrows}
\title{Move Arrows}
\usage{
tarrows(x0, y0, x1, y1, x2 = x0, y2 = y0, x3 = x1, y3 = y1, t, when,
  p = 1, length = 0.125, ...)
}
\arguments{
\item{x0, x1}{numeric vectors, start arrow defining x coordinates.}

\item{y0, y1}{numeric vector, start arrow defining y coordinates.}

\item{x2, x3}{numeric vector, end arrow defining x coordinates.}

\item{y2, y3}{numeric vector, end arrow defining y coordinates.}

\item{t}{numeric, homotopy parameter, limited between 0 and 1.
This parameter can be considered as fraction of animation duration time.}

\item{when}{numeric vector. This parameter controls
the times of: entrance, exit, start of movement and, end of movement.}

\item{p}{numeric, homotopy power parameter. Defaults to 1.}

\item{length}{length of the edges of the arrow head (in inches).}

\item{...}{additional arguments passed to \code{\link{arrows}}.}
}
\value{
List with numerical components \code{x} and \code{y} with
  current position
}
\description{
Move arrows from start to end location
}
\examples{
x0 <- c(0,4,7)
y0 <- c(0,7,0)
x1 <- c(4,0,7)
y1 <- c(4,7,5)
x2 <- c(0,4,7)+2
y2 <- c(0,7,0)-2
x3 <- c(4,0,7)+2
y3 <- c(4,7,5)-2
newplot()
arrows(x0,y0,x1,y1,lty=2)
arrows(x2,y2,x3,y3,lty=2)
pos <- tsegments(x0,y0,x1,y1,x2,y2,x3,y3,0.75)
pos
points(pos$start)
points(pos$end)
points(x0,y0,pch="0")
points(x1,y1,pch="1")
points(x2,y2,pch="2")
points(x3,y3,pch="3")
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\seealso{
\code{\link[graphics]{arrows}}, \code{\link{darrows}}
  for drawing arrows and \code{\link{tsegments}} for moving segments.
}
\keyword{aplot}
\keyword{dynamic}
\keyword{graphics}

