% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{tmatrix}
\alias{tmatrix}
\title{Interpolate Matrix}
\usage{
tmatrix(x0, x1 = diag(nrow(x0)), t, when, p = 1)
}
\arguments{
\item{x0}{numerical matrix, start matrix}

\item{x1}{numerical matrix, end matrix}

\item{t}{numeric, homotopy parameter, limited between 0 and 1.
This parameter can be considered as fraction of animation duration time.}

\item{when}{numeric vector. This parameter controls
the times of: entrance, exit, start of movement and, end of movement.}

\item{p}{numeric, homotopy power parameter. Defaults to 1.}
}
\value{
Intermediate matrix (numerical).
}
\description{
Interpolates matrix elements between start and end value.
}
\examples{
x0 <- matrix(c(1,2,3,4),2,2)
x1 <- matrix(c(11,12,13,14),2,2)
print(x0)
print(x1)
print(h(x0,x1,t=0.3))
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}

