% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatoR-functions-knitr.R
\name{tsegments}
\alias{tsegments}
\title{Move Segments}
\usage{
tsegments(x0, y0, x1, y1, x2 = x0, y2 = y0, x3 = x1, y3 = y1, t, when,
  p = 1, fixed = 0, ...)
}
\arguments{
\item{x0}{numeric vector, start x coordinates.}

\item{y0}{numeric vector, start y coordinates.}

\item{x1}{numeric vector, end x coordinates.}

\item{y1}{numeric vector, end y coordinates.}

\item{t}{numeric, homotopy parameter, limited between 0 and 1.
This parameter can be considered as fraction of time
duration of the animation.}

\item{when}{numeric vector. This parameter controls
the times of: entrance, exit, start of movement and, end of movement.}

\item{p}{numeric, homotopy power parameter. Defaults to 1.}

\item{fixed}{numeric, which location is fixed:
start (0, default) - draw from \code{(x0,y0)} to \code{(x1,y1)},
or end (1) - reverse.}

\item{...}{other parameters passed to \code{\link{segments}}.}
}
\value{
Numerical matrix with colums defining the current segments.
}
\description{
Change segment defining positions and plot subsequent segments.
Lines are "floating" to final positions.
}
\examples{
x0 <- c(0,4,7)
y0 <- c(0,7,0)
x1 <- c(4,0,7)
y1 <- c(4,7,5)
x2 <- c(0,4,7)+2
y2 <- c(0,7,0)-2
x3 <- c(4,0,7)+2
y3 <- c(4,7,5)-2
newplot()
arrows(x0,y0,x1,y1,lty=2)
arrows(x2,y2,x3,y3,lty=2)
pos <- tsegments(x0,y0,x1,y1,x2,y2,x3,y3,0.75)
pos
points(pos$start)
points(pos$end)
points(x0,y0,pch="0")
points(x1,y1,pch="1")
points(x2,y2,pch="2")
points(x3,y3,pch="3")
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\seealso{
\code{\link[graphics]{segments}}
}
\keyword{graphics}

